+++ server/dataAccess/models/user.js
@@ -1,0 +1,20 @@
+import mongoose from 'mongoose';
+const Schema = mongoose.Schema;
+
+const userSchema = new Schema({
+  username: {
+    type: String,
+    required: true,
+    index: {
+      unique: true,
+    },
+  },
+  password: {
+    type: String,
+    required: true,
+  },
+});
+
+const User = mongoose.model('User', userSchema);
+
+export default User;